move parse_E_layer to inchi module
